pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "stance-detection"
version = "0.1.0"
description = "U4RASD Stance detection code for the StanceNakba 2026 shared task"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.4.0",
    "transformers>=4.49.0",
    "accelerate>=0.26.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "pytorch-metric-learning>=2.3.0",
    "arabic-reshaper>=3.0.0",
    "python-bidi>=0.4.2",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[tool.setuptools.packages.find]
include = ["stance_detection", "stance_detection.*"]
